Made windows ssh path the default
In Windows 10 the powershell version is currently set to 5.1
which does not define these OS environment variables:

  $IsLinux
  $IsMacOS
  $IsWindows

This causes the executable ssh path to use the Linux/MacOS ssh path.
Mad windows ssh path the default so if the OS variables are not
defined, as in Windows 10 powershell version 5.1 it will use the
windows ssh path.